% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/A.R
\name{A}
\alias{A}
\title{An A function}
\usage{
A(h, f)
}
\arguments{
\item{h}{Hit rate.}

\item{f}{False alarm rate.}
}
\description{
This function allows you to calculate A from the hit rate and false alarm rate
}
\examples{
A(0.50, 0.25)
}
\keyword{A}

