% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/aPrime.R
\name{aPrime}
\alias{aPrime}
\title{An A prime function}
\usage{
aPrime(h, f)
}
\arguments{
\item{h}{Hit rate.}

\item{f}{False alarm rate.}
}
\description{
This function allows you to calculate A prime from the hit rate and false alarm rate
}
\examples{
aPrime(0.50, 0.25)
}
\keyword{A}
\keyword{prime}

