% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/b.R
\name{b}
\alias{b}
\title{A b function}
\usage{
b(h, f)
}
\arguments{
\item{h}{Hit rate.}

\item{f}{False alarm rate.}
}
\description{
This function allows you to calculate b from the hit rate and false alarm rate
}
\examples{
b(0.50, 0.25)
}
\keyword{b}
\keyword{prime}

