% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/bPrimed.R
\name{bPrimed}
\alias{bPrimed}
\title{A b double prime d function}
\usage{
bPrimed(h, f)
}
\arguments{
\item{h}{Hit rate.}

\item{f}{False alarm rate.}
}
\description{
This function allows you to calculate b double prime d from the hit rate and false alarm rate
}
\examples{
bPrimed(0.50, 0.25)
}
\keyword{b}
\keyword{d}
\keyword{double}
\keyword{prime}

