% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_over.R
\name{cross_over}
\alias{cross_over}
\title{A cross over function}
\usage{
cross_over(mod, cont_pred, grouping_var = FALSE, int_adj, slope_adj)
}
\arguments{
\item{mod}{A glm object}

\item{cont_pred}{The continuous (x) predictor}

\item{grouping_var}{(logical) Set to TRUE to include grouping var}

\item{int_adj}{With grouping variable, include intercept adjustment}

\item{slope_adj}{With grouping variable, include interaction (slope adjustment)}
}
\description{
This function allows you to calculate boundary/crossover point in 
logistic regression models.
}
\examples{
# Generate data
set.seed(1)
vot = rnorm(20, 15, 5)
vot = sort(vot)
phon1 = c(0,1,0,0,0,0,0,1,0,1,0,1,0,1,1,1,1,1,1,1)
group1 = rep('g1', 20)
df1 = data.frame(vot = vot, phon = phon1, group = group1)

phon2 = c(1,0,0,0,0,0,0,0,0,0,0,1,0,1,1,1,1,1,1,1)
group2 = rep('g2', 20)
df2 = data.frame(vot = vot, phon = phon2, group = group2)
df <- rbind(df1, df2)

# Fit model
glm1 <- glm(phon ~ vot, data = df, family = "binomial")
glm2 <- glm(phon ~ vot * group, data = df, family = "binomial")

# Get crossover point
cross_over(mod = glm1, cont_pred = 'vot')

# Get crossover point of grouping variables
cross_over(mod = glm2, cont_pred = 'vot')
cross_over(mod = glm2, cont_pred = 'vot', grouping_var = TRUE, 
           int_adj = 'groupg2', slope_adj = 'vot:groupg2')

# Plot regression with crossover point
library(ggplot2)
ggplot(df, aes(x = vot, y = phon)) + 
  geom_smooth(method = 'glm', method.args = list(family = 'binomial')) + 
  geom_vline(xintercept = cross_over(mod = glm1, cont_pred = 'vot'))

ggplot(df, aes(x = vot, y = phon, color = group)) + 
  geom_smooth(method = 'glm', method.args = list(family = 'binomial'), 
              se = FALSE) + 
  geom_vline(xintercept = cross_over(mod = glm2, cont_pred = 'vot')) +
  geom_vline(xintercept = cross_over(mod = glm2, cont_pred = 'vot', 
                                     grouping_var = TRUE, 
                                     int_adj = 'groupg2', 
                                     slope_adj = 'vot:groupg2'))
}
\keyword{crossover}
