% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/crossOver.R
\name{crossOver}
\alias{crossOver}
\title{A cross over function}
\usage{
crossOver(x)
}
\arguments{
\item{x}{A glm object}
}
\description{
This function allows you to calculate boundary/crossover point in
2AFC experiments.
}
\examples{
# Generate data
set.seed(1)
vot = rnorm(20, 15, 5)
vot = sort(vot)
phon = c(0,1,0,0,0,0,0,1,0,1,0,1,0,1,1,1,1,1,1,1)
df = as.data.frame(cbind(vot, phon))

# Fit model
glm <- glm(phon ~ vot, data = df, family = "binomial")

# Get crossover point
print(crossOver(glm))

# Plot regression with crossover point
plot(df$vot, df$phon, xlab = "vot", ylab = "phon", pch = 16, col = rgb(0, 0, 204, 102, maxColorValue = 255))
curve(predict(glm, data.frame(vot = x), type = "resp"), add = TRUE)
points(vot, fitted(glm), pch = 20)
abline(v = crossOver(glm), lty = 2, lwd = 0.75)
abline(h = 0.5, v = 0)
}
\keyword{crossover}

