% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distFeat.R
\docType{data}
\name{distFeat}
\alias{distFeat}
\title{Distinctive features of Spanish allphones}
\format{A data frame with 51 rows and 26 variables:}
\source{
\url{http://www.jvcasillas.com/}
}
\usage{
distFeat
}
\description{
A data frame containing +/-/0 specifications for the distinctive 
features of Spanish allophones. Based on the table by Casillas 
& Bessett (2013) published in Cedeño, R. A. N., Colina, S., & 
Bradley, T. G. (Eds.). Fonología generativa contemporánea de la 
lengua española. Georgetown University Press.
}
\keyword{datasets}

