% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distFeatDefs.R
\docType{data}
\name{distFeatDefs}
\alias{distFeatDefs}
\title{Definitions of distinctive features}
\format{A data frame with 51 rows and 26 variables:}
\source{
\url{http://www.jvcasillas.com/}
}
\usage{
distFeatDefs
}
\description{
A data frame containing definitions for the distinctive 
features of Spanish allophones. Supplement to the table by Casillas 
& Bessett (2013) published in Cedeño, R. A. N., Colina, S., & 
Bradley, T. G. (Eds.). Fonología generativa contemporánea de la 
lengua española. Georgetown University Press.
}
\keyword{datasets}

