% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/dPrime.R
\name{dPrime}
\alias{dPrime}
\title{A d prime function}
\usage{
dPrime(h, f)
}
\arguments{
\item{f}{Hit rate.}

\item{f}{False alarm rate.}
}
\description{
This function will calculate d prime. This is common in discrimination experiments.
Note: If your participants are at ceiling, you may want to consider another analysis.
}
\examples{
dPrime(.75, .25)
}
\keyword{d}
\keyword{prime}

