% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/eb.R
\name{eb}
\alias{eb}
\title{A function for adding error bars to barplots}
\usage{
eb(dv, fac, ...)
}
\arguments{
\item{dv}{Numeric vector.}

\item{fac}{Factors.}

\item{...}{Elipsis for passing arguements}
}
\description{
This function is a wrap around for barplot() that will add error bars. This
code was written by Dr. Mike Hammond at the U of A.
}
\examples{
eb(mtcars$mpg, as.factor(mtcars$cyl), ylab = "Miles per gallon", xlab = "cyl")
}
\keyword{bars}
\keyword{error}

