% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eucDist.R
\name{euc.dist}
\alias{euc.dist}
\title{A euclidean distance function}
\usage{
euc.dist(x1, x2)
}
\arguments{
\item{x1}{Numeric vector (generally f1).}

\item{x2}{Numeric vectors (generally f2).}
}
\description{
This function allows you to calculate the euclidean distance between two vectors
}
\examples{
euc.dist(rnorm(10, 10, 2), rnorm(10, 20, 2))
}
\keyword{distance}
\keyword{euclidean}

