% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euc_dist.R
\name{euc_dist}
\alias{euc_dist}
\title{A euclidean distance function}
\usage{
euc_dist(data, var1, var2)
}
\arguments{
\item{data}{A data frame}

\item{var1}{Numeric vector (generally f1).}

\item{var2}{Numeric vectors (generally f2).}
}
\description{
This function allows you to calculate the euclidean distance between two vectors
}
\examples{
# Create some data
df <- data.frame(vowel = gl(n = 2, k = 5, labels = c('a', 'o')),
                 var1 = rnorm(10, 10, 2), 
                 var2 = rnorm(10, 20, 2))

# Calculate euc.dist on entire data frame
euc_dist(data = df, var1 = var1, var2 = var2)

# Calculate euc.dist for each vowel
library(dplyr)
df \%>\%
 group_by(., vowel) \%>\%
 summarise(euc = euc_dist(., var1 = var1, var2 = var2))
}
\keyword{distance}
\keyword{euclidean}
