% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_unordered_list.R
\name{gen_unordered_list}
\alias{gen_unordered_list}
\title{An function to generate unordered lists from the directory structure}
\usage{
gen_unordered_list(path = ".", remove = NULL)
}
\arguments{
\item{path}{The path to the directory of which you want to print the 
structure}

\item{remove}{A regular expression to subset the output}
}
\description{
This function will create an unordered list of all the files in a directory 
and print the output. It should be used in conjunction with RMarkdown and 
knitr. Specifically, you want to call it within an R chunk that has the 
results set to 'asis'.
}
\examples{
# In any directory
gen_unordered_list()

# To specify the path
gen_unordered_list(path = "./")

# All files that are not .txt
gen_unordered_list(remove = ".txt")
}
\keyword{list}
\keyword{unordered}
