% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_logit.R
\name{inv_logit}
\alias{inv_logit}
\title{An inverse logit function}
\usage{
inv_logit(mod)
}
\arguments{
\item{mod}{A glm object}
}
\description{
This function allows you to calculate probability from log odds
}
\examples{
# Generate data
set.seed(1)
vot <- rnorm(20, 15, 5)
vot <- sort(vot)
fac <- rnorm(20, 100, 100)
phon <- c(0,1,0,0,0,0,0,1,0,1,0,1,0,1,1,1,1,1,1,1)
df <- data.frame(vot = vot, fac = fac, phon = phon)

# Fit models
glm0 <- glm(phon ~ vot, data = df, family = "binomial")
glm1 <- glm(phon ~ vot + fac, data = df, family = "binomial")
glm2 <- glm(phon ~ vot * fac, data = df, family = "binomial")
testLM <- lm(speed ~ dist, data = cars)

# Get beta weights as probabilities
library(dplyr)
inv_logit(glm0)
inv_logit(glm1)
inv_logit(glm2)
#inv_logit(testLM) # Gives an error
}
\keyword{inverse}
\keyword{logit}
