% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_logit_manual.R
\name{inv_logit_manual}
\alias{inv_logit_manual}
\title{An inverse logit function}
\usage{
inv_logit_manual(x)
}
\arguments{
\item{x}{An integer}
}
\description{
This function allows you to calculate probability from log odds
}
\examples{
inv_logit_manual(1)
}
\keyword{inverse}
\keyword{logit}
