% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_to_bib.R
\name{save_to_bib}
\alias{save_to_bib}
\title{A function for saving R packages to a .bib file}
\usage{
save_to_bib(package, file_name = NULL)
}
\arguments{
\item{package}{A character vector of the package you want to cite.}

\item{file_name}{A character vector for the file name of the output. Defaults to the package name.}
}
\description{
This function will generate a .bib file for an R package.
}
\examples{
# Create a .bib file called "test.bib" to cite the tidyverse
# save_to_bib("tidyverse", "test")
}
\keyword{bibtex}
