% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vowel_plot.R
\name{v_plot}
\alias{v_plot}
\title{A vowel plot function}
\usage{
v_plot(data, vowel, f1, f2, group = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{vowel}{A factor with vowels as levels}

\item{f1}{A continuous variable}

\item{f2}{A continuous variable}

\item{group}{An optional grouping variable for facets}
}
\description{
This function will plot f1/f2 data. The function takes 4 
obligatory arguments and optionally one can include a 
grouping variable ('group'). This function is a wrapper 
for ggplot. Columns from the data frame must be wrapped 
in quotation marks.
}
\examples{
# Vowel plot without grouping variable
v_plot(data = vowel_data, f1 = 'f1', f2 = 'f2', vowel = 'vowel', group = NULL)

# Vowel plot with grouping variable
v_plot(data = vowel_data, f1 = 'f1', f2 = 'f2', vowel = 'vowel', group = 'gender')
}
\keyword{plot}
\keyword{vowel}
