% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vowel_data.R
\docType{data}
\name{vowel_data}
\alias{vowel_data}
\title{Vowel data frame}
\format{A data frame with 750 rows and 6 variables:
\describe{
  \item{participant}{subject id labels}
  \item{gender}{male/female}
  \item{vowel}{Spanish oral vowels, i/e/a/o/u}
  \item{repitition}{iteration of item}
  \item{f1}{f1 values, in hz}
  \item{f2}{f2 values, in hz}
}}
\source{
\url{http://www.jvcasillas.com/}
}
\usage{
vowel_data
}
\description{
A dataset containing raw f1/f2 data based on data from Bradlow (1995)
and extended to include 50 participants
}
\keyword{datasets}
