% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b.R
\name{b}
\alias{b}
\title{A b function}
\usage{
b(data, h, f)
}
\arguments{
\item{data}{A data frame.}

\item{h}{A vector of hits (0 = miss, 1 = hit).}

\item{f}{A vector of false alarms (0 = correct rejection, 1 = false alarm).}
}
\description{
This function allows you to calculate b from a vector hits 
and a vector a false alarms.
}
\details{
Note: If you need to subset (by subject or by group), use 
b_() in conjunction with dplyr.
}
\examples{
# Create some data
set.seed(1); library(dplyr)
axb <- data.frame(subj = sort(rep(1:10, each = 20, times = 10)),
                  group = gl(2, 1000, labels = c("g1", "g2")),
                  hit = c(rbinom(1000, size = c(0, 1), prob = .8), 
                          rbinom(1000, size = c(0, 1), prob = .6)),
                  fa =  c(rbinom(1000, size = c(0, 1), prob = .3), 
                          rbinom(1000, size = c(0, 1), prob = .4))
)

# Calculate b on entire data frame
b(axb, hit, fa)

# Calculate b for each subject
# by group, plot it, and run a 
# linear model
axb \%>\%
  group_by(subj, group) \%>\%
  summarize(b = b_(., hit, fa)) \%T>\%
 {
  plot(b ~ as.numeric(group), data = ., 
       main = "b as a function of group", xaxt = "n", 
       xlab = "Group", ylab = "b")
  axis(1, at = 1:2, labels = c("g1", "g2"))
  abline(lm(b ~ as.numeric(group), data = .), col = "red")
 } \%>\%
 lm(b ~ group, data = .) \%>\%
 summary()
}
\keyword{b}
\keyword{prime}

