% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biVarPlot.R
\name{biVarPlot}
\alias{biVarPlot}
\title{A bivariate regression function}
\usage{
biVarPlot(data, x, y)
}
\arguments{
\item{data}{A data frame.}

\item{x}{An independent variable}

\item{y}{A dependent variable}
}
\description{
This function will fit a linear model to two continuous 
variables and plot the outcome.
}
\examples{
biVarPlot(data = cars, x = dist, y = speed)
biVarPlot(data = mtcars, x = cyl, y = mpg)
biVarPlot(data = Orange, x = age, y = circumference)
biVarPlot(data = ChickWeight, x = Time, y = weight)
biVarPlot(data = USArrests, x = UrbanPop, y = Rape)
}
\keyword{bivariate}
\keyword{regression}

