% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distFeat.R
\docType{data}
\name{distFeat}
\alias{distFeat}
\title{Distinctive features of Spanish allphones}
\format{A list of two dataframes.}
\source{
\url{http://www.jvcasillas.com/}
}
\usage{
distFeat
}
\description{
A list of two dataframes containing +/-/0 specifications for the 
distinctive features of Spanish allophones and definitions of the 
distinctice features. Based on the table by Casillas 
& Bessett (2013) published in Cedeño, R. A. N., Colina, S., & 
Bradley, T. G. (Eds.). Fonología generativa contemporánea de la 
lengua española. Georgetown University Press.
}
\keyword{datasets}

