% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dPrime.R
\name{dPrime}
\alias{dPrime}
\title{Function for calculating d prime}
\usage{
dPrime(data, h, f)
}
\arguments{
\item{data}{A data frame.}

\item{h}{A vector of hits (0 = miss, 1 = hit).}

\item{f}{A vector of false alarms (0 = correct rejection, 1 = false alarm).}
}
\description{
This function will calculate d prime from a vector of hits 
and a vector of false alarms.
}
\details{
This metric is common in discrimination experiments. 
Note: If you need to subset (by subject or by group), use 
dPrime_() in conjunction with dplyr. If your participants are
at ceiling, you may want to consider another analysis.
}
\examples{
# Create some data
set.seed(1); library(dplyr)
axb <- data.frame(subj = sort(rep(1:10, each = 20, times = 10)),
                  group = gl(2, 1000, labels = c("g1", "g2")),
                  hit = c(rbinom(1000, size = c(0, 1), prob = .8), 
                          rbinom(1000, size = c(0, 1), prob = .6)),
                  fa =  c(rbinom(1000, size = c(0, 1), prob = .3), 
                          rbinom(1000, size = c(0, 1), prob = .4))
)

# Calculate d prime on entire data frame
dPrime(axb, hit, fa)


# Calculate d prime for each subject by group, plot it, 
# and run a linear model
axb \%>\%
  group_by(subj, group) \%>\%
  summarize(dp = dPrime_(., hit, fa)) \%T>\%
 {
  plot(dp ~ as.numeric(group), data = ., 
       main = "d' as a function of group", xaxt = "n", 
       xlab = "Group", ylab = "d' prime")
  axis(1, at = 1:2, labels = c("g1", "g2"))
  abline(lm(dp ~ as.numeric(group), data = .), col = "red")
 } \%>\%
 lm(dp ~ group, data = .) \%>\%
 summary()
}
\keyword{d}
\keyword{prime}

